requirements/base.txt
django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
django-cors-headers==4.9.0
psycopg[binary]==3.3.3
python-dotenv==1.1.1
